Implement amalgamation of C modules.
C modules are embedded into the amalgamated script as a string
and written out to a temporary file on demand. The temporary file
is then loaded via package.loadlib.
For safety, portability, and compatibility, C module embedding is
only activated when the -x flag is specified.
Also update documentation and test cases.