Try to cleanup temporary DLL files.
A sentinel value with a __gc metamethod is stored away for every
temporary library file that's created. The __gc metamethod tries
to remove the library file, but on some OS/Lua version combinations
this may not work because it would try to remove a file that's still
in use (loaded as a library).